@@ -4,7 +4,7 @@ xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/ma
    <parent>
 	   <groupId>org.springframework.boot</groupId>
 	   <artifactId>spring-boot-starter-parent</artifactId>
-	   <version>2.7.5</version>
+	   <version>2.7.8</version>
 	   <relativePath />
    </parent>
 
@@ -17,10 +17,9 @@ xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/ma
 	<inceptionYear>2022</inceptionYear>
 
 	<properties>
-        <java.version>11</java.version>
+        <java.version>17</java.version>
 		<spring-cloud.version>2021.0.5</spring-cloud.version>
-		<feign-reactor.version>3.1.1</feign-reactor.version>
-		<snakeyaml.version>1.32</snakeyaml.version>
+		<snakeyaml.version>1.33</snakeyaml.version>
 	</properties>
 	<dependencyManagement>
 		<dependencies>
@@ -36,7 +35,7 @@ xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/ma
 	<dependencies>
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-starter-kubernetes-fabric8</artifactId>
+			<artifactId>spring-cloud-starter-kubernetes-fabric8-config</artifactId>
 		</dependency>	
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
