@@ -5,7 +5,7 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.7.5</version>
+		<version>3.0.2</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.example</groupId>
@@ -14,9 +14,8 @@
 	<name>where-for-dinner-notify</name>
 	<description>Where for Dinner notification application</description>
 	<properties>
-		<java.version>11</java.version>
-		<spring-cloud.version>2021.0.5</spring-cloud.version>
-		<snakeyaml.version>1.32</snakeyaml.version>
+		<java.version>17</java.version>
+		<spring-cloud.version>2022.0.0</spring-cloud.version>
 	</properties>
 	<dependencies>
 		<dependency>
@@ -63,25 +62,6 @@
 			</dependency>
 		</dependencies>
 	</dependencyManagement>
-    <repositories>
-        <repository>
-            <id>spring-releases</id>
-            <url>https://repo.spring.io/release/</url>
-            <releases>
-                <enabled>true</enabled>
-            </releases>
-            <snapshots>
-                <enabled>false</enabled>
-            </snapshots>
-        </repository>
-    </repositories>
-	<pluginRepositories>
-	    <pluginRepository>
-	        <id>spring-release</id>
-	        <name>Spring release</name>
-	        <url>https://repo.spring.io/release</url>
-	    </pluginRepository>
-	</pluginRepositories>
     <build>
         <plugins>
 			<plugin>
@@ -112,13 +92,6 @@
 			   <dependency>
 					<groupId>org.springframework.cloud</groupId>
 					<artifactId>spring-cloud-stream-binder-rabbit</artifactId>
-			   </dependency>
-			   <dependency>
-					<groupId>org.springframework.cloud</groupId>
-					<artifactId>spring-cloud-stream</artifactId>
-					<scope>test</scope>
-					<classifier>test-binder</classifier>
-					<type>test-jar</type>
 			   </dependency>
 		   </dependencies>
        </profile>
