@@ -4,8 +4,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Primary;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
+import org.springframework.util.StringUtils;
 
-import com.java.example.tanzu.wherefordinner.feign.CrawlerClient;
+import com.java.example.tanzu.wherefordinner.exchange.CrawlerClient;
 import com.java.example.tanzu.wherefordinner.model.Availability;
 import com.java.example.tanzu.wherefordinner.model.SearchCriteria;
 import com.java.example.tanzu.wherefordinner.searcher.Searcher;
@@ -28,7 +29,9 @@ public class CrawlerSearcher implements Searcher
 	{
 		log.info("Making crawler dining search for dining search {}", crit.getName());
 		
-		return crawlClient.search(crit.getDiningNames(), crit.getDiningTypes(), crit.getStartTime(), crit.getEndTime())
+		return crawlClient.search(StringUtils.hasText(crit.getDiningNames()) ? crit.getDiningNames() : "", 
+				StringUtils.hasText(crit.getDiningTypes()) ? crit.getDiningTypes() : "", 
+				crit.getStartTime(), crit.getEndTime())
 			.map(avail -> 
 			{
 				avail.setSearchName(crit.getName());
